requirements.txt
tkinter
pyperclip
PIL
azure.cognitiveservices.vision.computervision
msrest